move file/unarchive to unarchive resource
