+++ load/resource.go
@@ -35,7 +35,6 @@ import (
 	_ "github.com/asteris-llc/converge/resource/file/fetch"
 	_ "github.com/asteris-llc/converge/resource/file/mode"
 	_ "github.com/asteris-llc/converge/resource/file/owner"
-	_ "github.com/asteris-llc/converge/resource/file/unarchive"
 	_ "github.com/asteris-llc/converge/resource/group"
 	_ "github.com/asteris-llc/converge/resource/lvm/fs"
 	_ "github.com/asteris-llc/converge/resource/lvm/lv"
@@ -46,6 +45,7 @@ import (
 	_ "github.com/asteris-llc/converge/resource/param"
 	_ "github.com/asteris-llc/converge/resource/shell"
 	_ "github.com/asteris-llc/converge/resource/shell/query"
+	_ "github.com/asteris-llc/converge/resource/unarchive"
 	_ "github.com/asteris-llc/converge/resource/user"
 	_ "github.com/asteris-llc/converge/resource/wait"
 	_ "github.com/asteris-llc/converge/resource/wait/port"
