+++ samples/unarchive.hcl
@@ -1,5 +1,5 @@
-# file unarchive
-file.unarchive "consul.zip" {
+# unarchive
+unarchive "consul.zip" {
   source      = "/tmp/consul.zip"
   destination = "/tmp/consul"
 }
